Upgrade npm dependencies, notably nan to 1.6.2
This allows v8-debug to build under io.js and Node 0.12. However, v8-debug does not run under io.js! There is a problem where `processDebugRequest` is not getting called and the test hangs. This isn't a regression and this commit is in the right direction to getting v8-debug and node-inspector working on io.js -- it needs more investigation though.